+++ structurizr-export/src/main/java/com/structurizr/export/plantuml/StructurizrPlantUMLExporter.java
@@ -428,6 +428,7 @@ public class StructurizrPlantUMLExporter extends AbstractPlantUMLExporter {
         );
         plantUMLStyles.add(plantUMLRelationshipStyle);
 
+        int metadataFontSize = calculateMetadataFontSize(style.getFontSize());
         String description = "";
         String technology = relationship.getTechnology();
 
@@ -451,13 +452,14 @@ public class StructurizrPlantUMLExporter extends AbstractPlantUMLExporter {
             }
 
             writer.writeLine(
-                    String.format("%s %s%s %s <<%s>> : %s",
+                    String.format("%s %s%s %s <<%s>> : %s%s",
                             idOf(relationship.getSource()),
                             arrowStart,
                             arrowEnd,
                             idOf(relationship.getDestination()),
                             plantUMLRelationshipStyle.getClassSelector(),
-                            description));
+                            description,
+                            (StringUtils.isNullOrEmpty(technology) ? "" : "\\n<size:" + metadataFontSize + ">[" + technology + "]</size>")));
         } else {
             String arrow;
 
@@ -466,12 +468,6 @@ public class StructurizrPlantUMLExporter extends AbstractPlantUMLExporter {
             } else {
                 arrow = "-->";
             }
-
-//            if (!isVisible(view, relationshipView)) {
-//                relationshipStyle = "hidden";
-//            }
-
-            int metadataFontSize = calculateMetadataFontSize(style.getFontSize());
 
             // 1 --> 2 : "...\n<size:..>...</size>
             writer.writeLine(format("%s %s %s <<%s>> : \"%s%s\"",